requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
ca-certificates=2019.11.27=0
certifi=2019.11.28=py37_0
cffi=1.13.2=py37h7a1dbc1_0
click=7.0=py37_0
cpuonly=1.0=0
flask=1.1.1=py_0
freetype=2.9.1=ha9979f8_1
icc_rt=2019.0.0=h0cc432a_1
intel-openmp=2019.4=245
itsdangerous=1.1.0=py37_0
jinja2=2.10.3=py_0
jpeg=9b=hb83a4c4_2
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_0
markupsafe=1.1.1=py37he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
ninja=1.9.0=py37h74a9793_0
numpy=1.17.4=py37h4320e6b_0
numpy-base=1.17.4=py37hc3f5095_0
olefile=0.46=py37_0
openssl=1.1.1d=he774522_3
pillow=6.2.1=py37hdc69c19_0
pip=19.3.1=py37_0
pycparser=2.19=py37_0
python=3.7.5=h8c8aaf0_0
pytorch=1.3.1=py3.7_cpu_0
setuptools=42.0.2=py37_0
six=1.13.0=py37_0
sqlite=3.30.1=he774522_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.4.2=py37_cpu
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
wincertstore=0.2=py37_0
xz=5.2.4=h2fa13f4_4
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
